requirements.txt
Flask>=2.3.3
requests>=2.31.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
